fix example issues
