+++ examples/type-to-search-wiki/README.md
@@ -4,7 +4,6 @@ Small vanilla js example for represent rubico functions:
 
 - `pipe`
 - `switchCase`
-- `reduce`
 - `noop`
 
 Installation
@@ -13,14 +12,8 @@ Installation
 npm install
 ```
 
-Run (dev)
+Run example (dev)
 
 ```bash
 npm start
-```
-
-Build
-
-```bash
-npm run build
 ```
